requirements.txt
openai-agents
youtube-transcript-api
python-dotenv
streamlit
werkzeug
